Revert "Merge pull request #93 from denisgindulin/keller/removeCoordinator"
This reverts commit 4b6223fe4b6eb97435cf0a0deda4405d8562361b, reversing
changes made to 26e99022f759f803a669e0243b27658dab7bc56d.

+++ EventHub/Services/CoreData/FavoriteEvent+CoreDataClass.swift
@@ -20,7 +20,7 @@ extension FavoriteEvent {
         return NSFetchRequest<FavoriteEvent>(entityName: "FavoriteEvent")
     }
 
-    @NSManaged public var id: Int
+    @NSManaged public var id: Int64
     @NSManaged public var image: String?
     @NSManaged public var title: String?
     @NSManaged public var date: Date?
